test: split oversized test modules by topic
editor/mode/tests.rs (1985 lines) and history/import/tests.rs (1422
lines) had grown large enough that duplicate cases could slip in
unnoticed. Turn each into a directory module with topic-focused
siblings. Tests move verbatim; no test was added, removed or renamed.
