+++ crates/arf-console/src/editor/mode/tests/mod.rs
@@ -1,0 +1,10 @@
+//! Tests for `editor::mode`, split by topic into sibling modules.
+
+mod auto_match;
+mod bracket_delete;
+mod buffer_state;
+mod closing_conditions;
+mod quote_auto_match;
+mod skip_over_rules;
+mod state_basics;
+mod tree_sitter_word_nav;
